selinux: show permissive state for system services
Due to the project nature of giving SELinux a big fat no,
we've been hacking kernel to show enforcing while internally
running as permissive. However, this is checked additionally
from the SELinux library from the userspace and it blocks many
operations that we want to actually work. Most recently,
it started to block critical system services to run and prevented boot.
To workaround this, report permissive to the userspace while spoofing GMS
and regular user apps to detect enforcing.